+++ src/exactly_lib/definitions/primitives/boolean.py
@@ -5,3 +5,10 @@ BOOLEANS = {
     False: FALSE,
     True: TRUE,
 }
+
+BOOLEANS_STRINGS = {
+    item[1]: item[0]
+    for item in BOOLEANS.items()
+}
+
+CONSTANT_MATCHER = 'constant'